+++ scrapli_paramiko/__init__.py
@@ -1,3 +1,3 @@
 """paramiko transport plugin for scrapli"""
 
-__version__ = "2020.06.06"
+__version__ = "2020.10.10"