+++ scrapli_paramiko/transport/miko.py
@@ -1,6 +1,5 @@
 """scrapli_paramiko.transport.miko"""
 from logging import getLogger
-from threading import Lock
 from typing import Optional, Tuple
 
 from paramiko import Channel
@@ -38,10 +37,6 @@ class MikoTransport(Transport):
         timeout_socket: int = 5,
         timeout_transport: int = 5,
         timeout_exit: bool = True,
-        keepalive: bool = False,
-        keepalive_interval: int = 30,
-        keepalive_type: str = "",
-        keepalive_pattern: str = "\005",
         ssh_config_file: str = "",
         ssh_known_hosts_file: str = "",
     ) -> None:
@@ -60,19 +55,7 @@ class MikoTransport(Transport):
             auth_strict_key: True/False to enforce strict key checking (default is True)
             timeout_socket: timeout for establishing socket in seconds
             timeout_transport: timeout for ssh transport in seconds
-            timeout_exit: True/False close transport if timeout encountered. If False and keepalives
-                are in use, keepalives will prevent program from exiting so you should be sure to
-                catch Timeout exceptions and handle them appropriately
-            keepalive: whether or not to try to keep session alive
-            keepalive_interval: interval to use for session keepalives
-            keepalive_type: network|standard -- 'network' sends actual characters over the
-                transport channel. This is useful for network-y type devices that may not support
-                'standard' keepalive mechanisms. 'standard' is not currently implemented w/ paramiko
-            keepalive_pattern: pattern to send to keep network channel alive. Default is
-                u'\005' which is equivalent to 'ctrl+e'. This pattern moves cursor to end of the
-                line which should be an innocuous pattern. This will only be entered *if* a lock
-                can be acquired. This is only applicable if using keepalives and if the keepalive
-                type is 'network'
+            timeout_exit: True/False close transport if timeout encountered
             ssh_config_file: string to path for ssh config file
             ssh_known_hosts_file: string to path for ssh known hosts file
 
@@ -94,10 +77,6 @@ class MikoTransport(Transport):
             timeout_socket,
             timeout_transport,
             timeout_exit,
-            keepalive,
-            keepalive_interval,
-            keepalive_type,
-            keepalive_pattern,
         )
 
         self.auth_username: str = auth_username or cfg_user
@@ -106,7 +85,6 @@ class MikoTransport(Transport):
         self.auth_strict_key: bool = auth_strict_key
         self.ssh_known_hosts_file: str = ssh_known_hosts_file
 
-        self.session_lock: Lock = Lock()
         self.session: ParamikoTransport
         self.channel: Channel
 
@@ -155,13 +133,12 @@ class MikoTransport(Transport):
         """
         if not self.socket.socket_isalive():
             self.socket.socket_open()
-        self.session_lock.acquire()
+
         try:
             self.session = ParamikoTransport(self.socket.sock)
             self.session.start_client()
         except Exception as exc:
             LOG.critical(f"Failed to complete handshake with host {self.host}; Exception: {exc}")
-            self.session_lock.release()
             raise exc
 
         if self.auth_strict_key:
@@ -172,13 +149,9 @@ class MikoTransport(Transport):
         if not self._isauthenticated():
             msg = f"Authentication to host {self.host} failed"
             LOG.critical(msg)
-            self.session_lock.release()
             raise ScrapliAuthenticationFailed(msg)
-        self._open_channel()
-        self.session_lock.release()
 
-        if self.keepalive:
-            self._session_keepalive()
+        self._open_channel()
 
     def _verify_key(self) -> None:
         """
@@ -346,11 +319,9 @@ class MikoTransport(Transport):
             N/A
 
         """
-        self.session_lock.acquire()
         self.channel.close()
         LOG.debug(f"Channel to host {self.host} closed")
         self.socket.socket_close()
-        self.session_lock.release()
 
     def isalive(self) -> bool:
         """
@@ -403,7 +374,7 @@ class MikoTransport(Transport):
         """
         self.channel.send(channel_input)
 
-    def set_timeout(self, timeout: Optional[int] = None) -> None:
+    def set_timeout(self, timeout: int) -> None:
         """
         Set session timeout
 
@@ -417,24 +388,4 @@ class MikoTransport(Transport):
             N/A
 
         """
-        if isinstance(timeout, int):
-            set_timeout = timeout
-        else:
-            set_timeout = self.timeout_transport
-        self.channel.settimeout(set_timeout)
-
-    def _keepalive_standard(self) -> None:
-        """
-        Send 'out of band' (protocol level) keepalives to devices.
-
-        Args:
-            N/A
-
-        Returns:
-            N/A  # noqa: DAR202
-
-        Raises:
-            NotImplementedError: always, because this is not implemented for paramiko transport
-
-        """
-        raise NotImplementedError("No 'standard' keepalive mechanism for paramiko.")
+        self.channel.settimeout(timeout)