+++ CHANGELOG.md
@@ -1,17 +1,21 @@
 CHANGELOG
 =======
 
-# 2020.06.06
+# 2020.10.10
+- Add 3.9 to actions
+- Update pins
+- Remove keepalive stuff for now (in line w/ scrapli core)
+- Remove transport session locks
 
+
+# 2020.06.06
 - Minor internal changes, changed to use nox, better linting/typing because we now snag the scrapli type stubs, fixed
  a few docstrings, mostly just housekeeping!
 
 
 # 2020.04.30
-
 - Minor change to handle ssh keys and keep in line with scrapli core.
 
 
 # 2020.04.18
-
 - Initial separation of paramiko transport into its own repo as a plugin